Use this tool on itself
To make development easier, I set up a template Excel file.
This Excel file is blank and has references to the nessesary
libraries set already. To build the Add-in, open the template
file and then import the files listed in the CodeExportFileList.conf
(perhaps using a version of this tool already installed).

Note that I have had to leave an empty module in the template file
because Excel will delete the library references without it.
This may be a problem to address with this tool.