Use VBA-JSON for config serialization
Previously, the format used for the config file was a simple
list of key-value pairs which was represented in a text file using
line breaks and colons. Additionally, square brackets had special
meaning as well. Even for this simple serialization requirement,
there was already a lot that the developer has to think about.
For example there was a function which cleaned out strings to make
sure the didn't break the serialization.

A feature which will be implemented soon would be a lot easier
to do if the config file was in a format such as JSON. Since
JSON is fairly easy to read and write it seems like it doesn't
introduce complexity. There is a popular open source VBA library
for reading and writing JSON called VBA-JSON. I used this library
and this simplified the data serialization code dramatically.

I changed a few details of how the actual importing and exporting
works, however there should be no loss of functionality.

Relative and absolute file paths should now be supported,
however I have not tested this yet. I have only tested the case
where all files are in the base directory.