Support importing library references
Library references (for example to the extensibility library) are
removed from an Excel file if the file is saved with no VBA code
in it. This may happen when using this add-in since all of the VBA
code is probably going to be exported before save. To solve this,
the add-in can now automatically inject references to libraries
when the user calls the import action. The details of the libraries
are declared in the configuration file. When the configuration
file is automatically generated, it will read the library references
which are currently referenced and list all the appropriate details
in the configuration file.

The "References" property of the configuration file is to be an
array of objects. Each object represents a library reference with
the following properties:

* Name - name of the library
* Description (optional) - description of the library
* GUID - the GUID of the library
* Major - the major version number of the library
* Minor - the minor version number of the library
* Path - The path to the library on the file system

If the GUID property is given then the GUID, Major and Minor
properties are used to locate the library, otherwise the Path
property is used.