Fix problems with importing Dictionary.cls
@spence10 was having trouble importing the Dictionary.cls. The
VBA-Dictionary project has encountered this same problem before
and concluded that the problem was due to having Unix style line
endings in the VBA source files (especially class modules).

Firstly, a git attributes file was set up to make sure that VBA
modules are converted to windows style line endings. Secondly,
the use of the drop in Dictionary replacement is so that Excel
for Mac is supported. However, the Microsoft scripting library
is already used for file system purposes and therefore Mac isn't
supported anyway. So there should be no loss of compatibility by
using the Microsoft scripting library instead of the Dictionary
drop in replacement. The Dictionary class was removed.