Fix issue #21
In the Export() sub, an unhandled error would occur if the
"References" or "Module Paths" properties were not set in the
JSON configuration file.

The problem was handled by making these properties optional. If
they are not present, the Export sub will not perform the action
related to each property. This means the behaviour for an empty
property is identical to no property. This is the behaviour already
implemented in the Import sub. I forgot to change the Export sub
when I did the Import implementation.