Allow debugging from error dialog
When an error occurs an error dialog will show as before. However
this dialog will ask if the user would like to enter the debugger.

The error handler is still only implemented in the same procedures
as before. Before putting it in all procedures I want to wait for
another PR to get merged so that I can rename the exitSub label
to exitProc.