Only add non-empty modules to configuration
This is a fix for an issue mainly with sheet modules. It is normal
for sheet modules not to have any code. In this case it is
distracting to have empty files in the project directory just for
these. Instead they shouldn't be added to the configuration file.

I hope this won't cause any problems with modules left intentionally
empty.