Modularise config logic & improve behaviour
The primary goal of this commit is to extract the configuration
logic from the importing and exporting logic. The configuration
logic is now in a class clsConfiguration. This class provides
an easy to use and safe API for working with the configuration.
The user of the API should be safe from causing corruption.
Additionally the user of the API should be shielded from low
level aspects of a corrupted configuration file. What I'm trying
to say is that it adds a layer of abstraction.

While porting the import and export logic to use the new API, I
improved some behaviour of the MakeConfigFile routine and the
Export routine.

In the MakeConfigFile, the routine now takes
advantage of the non-destructive nature of the configuration API.
The MakeConfigFile now updates the config file instead of creating
a new one. It will silently add modules and references but will
prompt the user to ask if missing modules and references should
be deleted.

In the Export, the routine now silently ignores modules listed
in the configuration file that isn't currently in the project.
This partially solves issue #30. The only thing left to do is
to remove and clear the modules after everything else has
successfully exported.

Some minor refactoring was done. The modUtil module was created
to house functionality that serves a very general purpose.

Comments were added to each function and module subsection to
help with documentation. Comment at the top of modImportExport
and modMenu describing the references nessesary for the code to
run was removed since an authoritive list of references required
is kept in the CodeExport.config.json.

.editorconfig was created to aid in editing the VB code and other
code in modern text editors.

Note: I am yet to test this code. It compiles without error but
may have runtime errors. I just want to get it committed to
make sure I don't loose work.